Converting static libraries to dynamic libraries
