+++ BitsoundPlayer.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module BitsoundPlayer {
+  umbrella header "BitsoundPlayer.h"
+
+  export *
+  module * { export * }
+}
