--- BitsoundReceiver.framework/Resources
@@ -1,1 +1,0 @@
-Versions/Current/Resources
