+++ BitsoundShaking.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module BitsoundShaking {
+  umbrella header "BitsoundShaking.h"
+
+  export *
+  module * { export * }
+}
